refactor(tasks): move pending_order_sync logic to jobs/ layer
Adapter classes and DB session management moved to
app/jobs/pending_order_sync. Task wrapper delegates to
run_pending_order_sync() per AGENTS.md convention.
